+++ src/PSForms.psm1
@@ -1,8 +1,8 @@
 # Private Functions
 . $PSScriptRoot\Private\Functions\ConvertTo-PSFormsPSHTML.ps1
-. $PSScriptRoot\Private\Functions\Get-PSFormsResult.ps1
 
 # Public Functions
+. $PSScriptRoot\Functions\Get-PSFormsResult.ps1
 . $PSScriptRoot\Functions\New-PSFormsSimpleForm.ps1
 . $PSScriptRoot\Functions\New-PSFormsSite.ps1
 . $PSScriptRoot\Functions\New-PSFormsInput.ps1
@@ -11,12 +11,11 @@
 . $PSScriptRoot\Functions\New-PSFormsParagraph.ps1
 
 
-
-
+# Exports
 Export-ModuleMember -Function New-PSFormsParagraph
 Export-ModuleMember -Function New-PSFormsInput
 Export-ModuleMember -Function New-PSFormsInputMultipleChoice
 Export-ModuleMember -Function New-PSFormsInputTextBox
 Export-ModuleMember -Function New-PSFormsSite
 Export-ModuleMember -Function New-PSFormsSimpleForm
-
+Export-ModuleMember -Function Get-PSFormsResult