+++ Changelog.md
@@ -1,9 +1,29 @@
 # Change Log
 
-All notable changes to this project will be documented in this file.
-
-The format is based on [Keep a Changelog](http://keepachangelog.com/)
-and this project adheres to [Semantic Versioning](http://semver.org/).
-
 ## [0.1.0] Unreleased
 
+- Initial Release
+
+### Added
+
+- Functions:
+    - New-PSFormsInput
+    - New-PSFormsInputMultipleChoice
+    - New-PSFormsInputTextBox
+    - New-PSFormsParagraph
+    - New-PSFormsSimpleForm
+    - New-PSFormsSite
+    - Get-PSFormsResult
+    - Private:
+        - ConvertTo-PSFormsPSHTML
+- Objects in C# lib:
+    - PSFormsParagraph
+    - PSFormsInputTextBox
+    - PSFormsInputMultiple
+    - PSFormsInput
+- New PSForms Site Template via Plaster Template
+- New HomePage Template
+- Simple Form Template
+- Tests:
+    - PowerShell Script Analyzer Running Against all Function Files
+    - Simple Module Tests